updated Van class

+++ Cameron_Jackson_Code/Week1/WheelinDealin/src/com/ex/vehicles/Vehicle.java
@@ -32,7 +32,7 @@ public abstract class Vehicle {
 		return this.passengerCapacity_;
 	}
 	
-	float loadCapacity() {
+	double loadCapacity() {
 		return 0;
 	}
